Move int64 <> byte conversion.
